+++ app/frontend/src/javascript/components/user/change-password.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next';
 import Authentication from '../../api/authentication';
 import { FieldValues } from 'react-hook-form/dist/types/fields';
 import { PasswordInput } from './password-input';
-import { FormState } from 'react-hook-form/dist/types/form';
+import { FormState, UseFormSetValue } from 'react-hook-form/dist/types/form';
 import MemberAPI from '../../api/member';
 import { User } from '../../models/user';
 
@@ -18,13 +18,14 @@ interface ChangePasswordProp<TFieldValues> {
   currentFormPassword: string,
   formState: FormState<TFieldValues>,
   user: User,
+  setValue: UseFormSetValue<User>,
 }
 
 /**
  * This component shows a button that trigger a modal dialog to verify the user's current password.
  * If the user's current password is correct, the modal dialog is closed and the button is replaced by a form to set the new password.
  */
-export const ChangePassword = <TFieldValues extends FieldValues>({ register, onError, currentFormPassword, formState, user }: ChangePasswordProp<TFieldValues>) => {
+export const ChangePassword = <TFieldValues extends FieldValues>({ register, onError, currentFormPassword, formState, user, setValue }: ChangePasswordProp<TFieldValues>) => {
   const { t } = useTranslation('shared');
 
   const [isModalOpen, setIsModalOpen] = React.useState<boolean>(false);
@@ -68,6 +69,7 @@ export const ChangePassword = <TFieldValues extends FieldValues>({ register, onE
     return handleSubmit((data: { password: string }) => {
       Authentication.verifyPassword(data.password).then(res => {
         if (res) {
+          setValue('current_password', data.password);
           setIsConfirmedPassword(true);
           toggleConfirmationModal();
         } else {
@@ -85,6 +87,7 @@ export const ChangePassword = <TFieldValues extends FieldValues>({ register, onE
         {t('app.shared.change_password.change_my_password')}
       </FabButton>}
       {isConfirmedPassword && <div className="password-fields">
+        <FormInput register={register} id="current_password" type="hidden" label="current password" />
         <PasswordInput register={register} currentFormPassword={currentFormPassword} formState={formState} />
       </div>}
       <FabModal isOpen={isModalOpen} toggleModal={toggleConfirmationModal} title={t('app.shared.change_password.change_my_password')} closeButton>