+++ app/models/user.rb
@@ -255,7 +255,7 @@ class User < ApplicationRecord
   end
 
   def password_complexity
-    return if password.blank? || SecurePassword.is_secured?(password)
+    return if password.blank? || SecurePassword.secured?(password)
 
     errors.add I18n.t('app.public.common.password_is_too_weak'), I18n.t('app.public.common.password_is_too_weak_explanations')
   end